Делит код на модули (util.js & data.js)
